backend/pyproject.toml
[build-system]
requires = ["uv_build>=0.7.5,<0.8"]
build-backend = "uv_build"

[project]
name = "backend"
version = "0.1.0"
description = "TanStack AI + pydantic-ai HITL Demo Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-ai",
    "pydantic-settings",
    "asyncpg",
    "duckdb",
    "pandas",
    "tanstack-pydantic-ai",
]

[tool.uv.sources]
tanstack-pydantic-ai = { git = "https://github.com/kzinmr/tanstack-pydantic-ai.git" }
